feat(build): add clean stage to build command
This commit adds a new stage to the build command which removes the
output directory before the project is built. Building a ts project will
overwrite existing files, but won't remove old ones. This was causing
old plugins which had since been deleted to stick around. This is now
fixed because the build directory is cleaned before the build occurs.